Add the core Plox class
_Crafting Interpreters_ begins by setting up a core Lox class to drive
the interpreter. It doesn't translate very nicely to Python where
there's no requirement to bundle functions up into classes. There's not
much going on here right now, so that would feel much more natural. But
the error tracking is just enough to make me just follow along and
hopefully not detach my foot right at the beginning.

This is not functional just yet (and neither is what's in the book.) We
need to get the scanner going first.