Add scanner with support for single char tokens
This is the very beginning of section 4.5, the start of the Scanner
class. Unfortunately we don't have a `switch` statement in python, so
we'll have to make do with a disgusting massive `if`/`else` block.

I'm not super down with the raw in memory string stuff we've got going
on here. I really want to open a stream object and run the scanner in a
pipeline. The standard library doesn't quite help us out here
unfortunately (I've already checked) so I'd have to implement that
myself. It would be worth it :P