Add comments to scanner
Looks like I'd actually borked this earlier. / characters need slightly
special handling in the scanner to check whether they're actually the
beginning of a comment. Which is to say they're _not_ a character that
uniquely defines a token.

So following the book I've broken that branch out and fixed it, and
modified the tests appropriately. Well, other way around actually.