Fix scanner to discard whitespace
The next step in the book is discarding whitespace characters. Rather
than add specific tests for this right now I've modified a couple of
existing hypothesis tests to include whitespace characters in their
generation. This needs a little bit of extra care, which is annoying but
it works.

These tests aren't particularly great, but they confirm that things are
working to some degree. I have some better ideas for more comprehensive
tests that I'll add in later.