Add an example Lox source scanner test
The unit-y tests that I have setup for the scanner are nice, and they
help to confirm some things. But it's a bit clunky trying to provide
longer form examples in that kind of setup. That's going to make it hard
to run larger tests of the interpreter's behaviour.

So I'm going to start putting example Lox scripts alongside some
simple serialisations of expected output from parts of the interpreter.
I've setup a pytest file that will search for these example files,
deserialise the intermediate output files and confirm that they match.
That should make it easy to write integration-y tests that actually
represent the expected outputs of the interpreter.