Add the beginnings of the parser
This brings us up to the end of chapter 6 (mostly...) by implementing
expressions in the parser. This translates pretty easily to python until
we get to the error handling, where we lack the function overloading of
Java and so I had to work around the reporting aspect.

I couldn't quite see how to test this until I got the translation
finished and poked it a bit. I added some very basic tests at that point
just to make sure we can recognise each part of the grammar. I haven't
tested that we generate the correct precedence or associativity yet. I'd
like to lean on hypothesis a bit here, because writing all the test
cases is boring :P But I haven't quite figured out how to do that nicely
yet, so I'll ponder it a bit more first.