Add some basic precedence and associativity tests
I added some tests to make sure that the parser could recognise each of
the expressions, but I didn't take care of the interactions between
them. This is kind of important, so I better add them now...

These are quite basic tests, but they should enumerate all of the
possible combinations of two expressions and assert that they're parsed
into the correct tree.