Add the initial interpreter implementation
Now that we can parse the syntax tree the next step is to pass that tree
off to a simple tree-walk interpreter. This is the first half of Chapter
7, where we optimistically implement the evaluation of the what the
current syntax tree can contain. That's limited to expressions, so we
basically have a functional calculator at this point.

It's an optimistic implementation because there's no error handling.
That appears to be the second half of the chapter.

This is reasonably covered by simplistic tests. I'm still being lazy and
not testing groupings, mostly because generating test cases feels like a
massive chore for that.