Add lox specific exceptions to the interpreter
Following the book I implemented the interpreter logic for evaluating
basic expressions, but that doesn't include any error handling to begin
with. First we need to raise some internal exception types to make the
handling that we'll add next more targeted.

I've only added what the book suggests here, but I think there are a
couple more cases that we could add now. It's possible that it will get
in the way of expanding the interpreter later on, but I think that could
be avoided (I'm looking at the ability to drop out of the main
expression switching logic in particular.) I'll follow the wisdom of the
book though so I don't end up painting myself into a corner.