Add the parser and interpreter to the main executor
I've had parts of this working for a while but I've not committed it.
Now that we have the parser and the interpreter running (to some extent
at least) we need to hook it up to our small executable framework (i.e.
the `main` function.) This is pretty straightforward. It also seems to
me more like an integration boundary, and that means I'm being lazy
about testing it.

Now that I look through the changes going in here I'm pretty sure some
of these were necessary for some of the recent changes, and so those
commits have actually been broken. Whoopsies!