Rename the source of the Expr class tree
I don't know why but I thought it would be better to call the Expr file
ast. I thought that it was being called Expr in the book purely to
satisfy Java, but now that I've read further I see that there is
actually also a separate Stmt class tree.