Refactor the generation script for AST classes
My implementation of the `generateAst` script is a bit different from
the one given in the book, because the generated python code is quite
different and python is easier to script with anyway. But I didn't have
the "foresight" that the book had and so my script didn't quite have the
flexibility needed to generate the statement classes (which is what
we're doing next.)

I've done much more than necessary here to add that flexibility in. I
think this looks a bit nicer (though perhaps a little harder for a
python neophyte to grok.) It could have been even better if it was just
a wrapper to render a jinja (or some other templating library) template,
but this version just uses the standard library and that seems nice for
this project.