Fix the import blocks in AST class files
The generation of the AST classes for statements has a subtle extra
requirement that I missed. It needs to import the `Expr` class, which
the generation script was not correctly importing. I've updated the
script to have a bit more flexibility with how it generates type hints,
and now it seems to work properly.

Clearly I've been bitten by not writing some tests. Nobody could have
predicted this!