Add new variable AST data structures
We need to add name binding, and the easiest place to start is with
global variables. First we need some extra data structures to transfer
the identifiers from the parser to the interpreter, and to shuffle the
grammar a bit in preparation for controlling the scoping rules.
Apparently, I'm just following what the book tells me here right now...

We're breaking the interpreter with this change, because it adds two new
abstract methods that we don't yet support.