Add global variable evaluation to interpreter
Now that the parser generates the necessary information we can fix the
interpreter by allowing it to work with global variables. We add an
Environment class to store the declared variables, and then wiring it
all up actually doesn't involve very much.

The environment class is really just a dictionary with a layer of
indirection in front. Maybe this is necessary in Java (I actually don't
know...) but it feels a little odd. It might be nicer to just subclass
`dict`, but then again it might not. And there may very well be some
good reason coming up, so I'll just follow along.