Add variable assignment expressions to the parser
I just added variable declaration and access, now we need assignment.
I've just followed the tricks given in the book here. I've only added
tests that assign literal values, not larger expressions. With the way
the parser works that should indicate that it does work correctly all
the way through the grammar, but it's not the most conclusive.

And we break the interpreter again...