Add a tree structure to environments
We add lexical scoping to the interpreter by generating a new
environment when we enter a new scope, and linking that environment to
the scope that is currently active. This will turn the Environment class
into a tree structure.

To make the scoping work we need to be able to climb up the tree in
order to find variables in the enclosing scopes. And for lexical scoping
we need to access and modify the innermost scope when possible.