Add support for block statements
The diff here is not too big, so I did it all in one hit :) Now that we
have the ability to support scoping we can add blocks to the parser to
make use of that. The interesting part is really done in the environment
here, so this commit is pretty straightforward. I vaguely expected the
environment shuffling to be more complicated, but I guess that's the
main point the book makes. Nothing in here is really _that_ magical.

One thing the book does is have the `block` method in the parser return
a list of statements rather than a `Block` statement object. This is so
that we can reuse the method elsewhere later. That feels a little odd to
me, especially when some very thin functions were introduced earlier.
But I'll go with it to make my life easier in the following chapters and
maybe refactor later.

Now that the language is getting more powerful it's becoming tedious to
setup test data to inject into the various components. So I've gone
pretty thin on the tests with this change. I'll either start introducing
more integration-y tests of source files, or I'll come up with a way to
more efficiently describe internal objects. But for now I'll just be
lazy :P