Add the final chapter 8 example to the test suite
At the end of chapter 8 there's a small Lox example that's meant to
demonstrate the statements added to the interpreter thus far. It seemed
like a nice example, so I've added it to the set (of one until now...)
and hooked up a new type of example test that checks the bytes written
to stdout.

I had to do a bit of refactoring of the `Token.from_string` method that
I used to get the example token tests working. There wasn't a
particularly compelling method for describing string literals using
that, so I've replace it with something else that is maybe a little
uglier. Oh well, it works.