Add support for binary logical operators
The other type of conditionals are short-circuiting binary logical
operators. It hadn't really occurred to me that these were a form of
control flow in the past...

I haven't included tests for side effects generated by these
expressions, again because it's becoming quite tedious to write out the
setup that I've been using.