Fix calling REPL after executing script
Looks like I'd nested a statement incorrectly in the orchestration
class, and that was causing the interpreter to enter a REPL after
executing a script.

I haven't been testing this part of the code, but I've started using the
test suite that's available in the book's GitHub repo and that seems to
be showing up an array of little mistakes.