Change the formatting in error messages
I'm trying to pass the tests that can be found in the github repository
for the book. They work by calling the executable and matching the
output on stdout, which means they're quite sensitive to the formatting
of the output messages. There's one part in particular that I hadn't
quite matched up, so this change brings it inline.