Fix parsing chained logical operators
Looks like I had not correctly implemented the substitution rules for
logical operators and was unable to parse chained logical operators.
I.e. something like
  a and b or c
This was because I'd stuffed a return statement into the while loop that
continued parsing the right operand, and so we climbed back up the
grammar immediately after parsing that operand instead of checking for
another boolean operator.