Add return statements to the parser and interpreter
This change almost brings us to the end of chapter 10 by adding return
statements to the syntax. The book uses exceptions to climb back up the
Java call stack, which translates exactly to python.

I feel like I'd prefer to catch the return exception in the interpreter
class, which is where it gets thrown. I feel like that would make it a
little easier to reason about the control flow. But I can see why the
book chooses to catch in the function call method, since it allows that
method to return the value. Perhaps I can find a refactoring later on
that reconciles these positions.