Add a variable resolver
To restore lexical scoping after adding closures we're going to add a
variable resolution stage in between the parser and the interpreter.
This commit adds the resolver itself, but doesn't yet hook it into
anything. This seems like a nice check-in point though.

I find some of the nomenclature in here a little confusing (particularly
`resolve`/`resolve_local`,) but I've followed the book so that I don't
make my life difficult in the future.