Add variable resolution to interpreter
After building the variable resolver we need to plumb it into the
interpreter. The translation from the book's Java code is not quite as
literal as it could be here; using python lists perhaps offer us a
little bit more flexibility than Java's stacks.

We have to remove the __eq__ method on the expression nodes that attrs
gives us. We need a hash method to be generated for the Variable nodes
so that we can track the resolution. Attrs won't allow such a method to
be generated when value comparisons are allowed because it violates
weird python internals. This is ok though because we should never need
these value comparisons.