Fix parsing of function call arguments
I'd somehow missed a bug in the parsing of function calls. The
implementation of the `call` method in the book uses a `do`/`while`
clause to parse the arguments. This type of statement doesn't exist in
python, and I'd kind of blindly translated it to a `while` statement.
That doesn't quite have the same behaviour though, since the
`do`/`while` will always execute its body at least once. To get the same
behaviour in python we need to place a copy of the body in front of the
loop (or perhaps use a more elegant for loop or something, this works
for the moment.)