Add the beginnings of the class machinery
The last two jlox chapters add classes to the language. At the
beginning of chapter 12 we add the basic syntax for declaring classes
and some data structures to stash everything. At this point we can parse
some very basic class declarations, but we don't bind any behaviour so
we can't actually achieve anything. It seems like a nice check in point
though.