Add super expressions
The last part of the jlox section adds super expressions to allow
a subclass to delegate to its superclass. The syntax here is a little
different to what I'm used to seeing, but I like it. It leads to a
slightly different implementation in the parser than I was expecting.

Some of the code here has ended up a bit ugly (in my opinion at least.)
I'm guessing it looks less so in Java, but I opted not to use some
python features that I could have to slim it down and that makes it feel
a little verbose.

But that's kind of it!