Fix a bug in error reporting
Looks like I had some problems creep in when I merged some overloaded
methods in the Java implementation into a single method with optional
arguments. This causes end of file token to not be formatted into error
messages.

It's caused by a typing issue, where I was comparing a token to a token
type. We actually have to shuffle this around conditional around a bit
to make sure that we handle this token being an optional argument.