Change LoxFunction to compare by identity
Because I'm using attrs everywhere we get free comparison by value
between object instances. However there is a test in the book's github
repository that enforces that functions are compared by identity. This
means that we should see
  > print foo.method == foo.method;
  false

I find this somewhat surprising, because it implies the following
  > var bound = foo.method;
  > var a = bound;
  > var b = bound;
  > print a == b;
  true
  > a = foo.method;
  > b = foo.method;
  > print a == b;
  false
and that doesn't sit too well with me. I would certainly find the
difference between those two sequences to be surprising. This is really
just an artifact of the parser, and I can vaguely imagine the
possibility of implementing in such a way that this couldn't really be
achieved.

But I'll take the book's tests as the language specification here, and
it's a trivial one line fix.