Fix default interpreter environment
I foolishly got bitten by the classic `attrs` default bug. There is a
small factory function to generate the global environment with the
(one) builtin names. This was being called to create the default value
assigned to the interpreter's environment. It is mutable though, so the
correct thing to do is to pass the factory function itself to `attrs`.

I don't think this is causing any problems for normal usage of the
interpreter right now, but it is affecting unit tests. And it's also
just bad.