Add PeekableTextIO
I'd like to start my refactoring adventures with the scanner, because
it's quite straightforward and didn't translate particularly well to
python. One thing I noticed that I thought could be improved was reading
the whole source into memory before scanning. It would be better if we
could operate on the stream directly.

We do still need some lookahead though, which the standard python stream
types don't provide. So before I get stuck into the scanner itself I've
pulled together a little buffer that I can wrap around another stream
object to get that lookahead.

A nice side bonus of this is that it provides another place for tracking
the line number. That's going to be helpful because I'm about to tear
the scanner apart and it'll save me from having to think about one more
thing.