Refactor scanner to use streams and dispatch
Big changeup time! Not so happy about that, but this is my lazy hobby
project and I wanted to make sure this was not a billion times slower
first.

This is a reimplementation of the scanner to
1. work with streaming text rather than an in memory string
2. use a form of dispatch instead of a big if/else statement

The original Java implementation in the book uses a switch statement to
select which token to scan. Python (somewhat infamously) doesn't have a
switch statement, so I had to translate that into an if/else statement
instead. I thought this looked a bit unpleasant, but conditionals are
also one of the bigger speed bumps in python.

So I've used a dictionary as a dispatch mechanism to select the scanners
behaviour. That's now fed by the PeekableTextIO class I added
previously. I think this does a better job of compartmentalising the
logic. It's also significantly faster, my profiling suggests that it's
about 50% faster.

This is still kind of messy, but it's currently working fairly nicely.
I'll try to come back to it at some point, clean up the names and look
into whether the individual scanning functions can be improved.