Add github actions CI
It turns out that I've broken a bunch of tests along the way after I
decided to start using the tests from the book's repository. Clearly
some CI is in order. I've been curious about Github actions so this
seems like a good opportunity to try it out a bit.